Port `load_amqp` and `save_amqp`
